+++ src/main/java/com/flightstats/hub/filter/EventsEncodingFilter.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.app;
+package com.flightstats.hub.filter;
 
 import com.google.common.base.Joiner;
 import org.apache.commons.lang3.StringUtils;
@@ -21,9 +21,9 @@ import java.util.List;
  * This exists to prevent gzip & deflate encodings for event-streams.
  */
 @PreMatching
-public final class StreamEncodingFilter implements ContainerResponseFilter {
+public final class EventsEncodingFilter implements ContainerResponseFilter {
 
-    private final static Logger logger = LoggerFactory.getLogger(StreamEncodingFilter.class);
+    private final static Logger logger = LoggerFactory.getLogger(EventsEncodingFilter.class);
 
     private List<String> removedEncodings = Arrays.asList(GzipFilter.GZIP, GzipFilter.DEFLATE);
 
@@ -48,17 +48,10 @@ public final class StreamEncodingFilter implements ContainerResponseFilter {
                         innerEncoding.add(token);
                     }
                 }
-                String joined = Joiner.on(",").join(innerEncoding);
-                if (StringUtils.isNotBlank(joined)) {
-                    allowedEncoding.add(joined);
-                }
+                allowedEncoding.add(Joiner.on(",").join(innerEncoding));
             }
             logger.debug("removing from events {} ", allowedEncoding);
-            if (allowedEncoding.isEmpty()) {
-                request.getHeaders().remove(HttpHeaders.ACCEPT_ENCODING);
-            } else {
-                request.getHeaders().put(HttpHeaders.ACCEPT_ENCODING, allowedEncoding);
-            }
+            request.getHeaders().put(HttpHeaders.ACCEPT_ENCODING, allowedEncoding);
         }
     }
